requirements.txt
pyqt6>=6.6.0
sounddevice>=0.4.6
numpy>=1.26.0
groq>=0.9.0
pywin32>=306
flask>=3.0.0
python-dotenv>=1.0.0
